+++ package.json
@@ -76,6 +76,7 @@
         "osmtogeojson": "^3.0.0-beta.4",
         "regenerator-runtime": "^0.13.7",
         "togpx": "nrenner/togpx#722d291",
+        "tokml": "^0.4.0",
         "topojson-client": "^3.1.0",
         "url-search-params": "~0.5.0"
     },
@@ -112,8 +113,8 @@
         "gulp-zip": "^5.0.2",
         "husky": "^4.3.4",
         "i18next-scanner": "^3.0.0",
-        "marked": "^2.0.0",
         "jest": "^26.6.3",
+        "marked": "^2.0.0",
         "merge-stream": "^2.0.0",
         "node-fetch": "^2.6.1",
         "npmfiles": "^0.1.3",
@@ -277,6 +278,11 @@
         "core-js-bundle": {
             "main": [
                 "minified.js"
+            ]
+        },
+        "tokml": {
+            "main": [
+                "tokml.js"
             ]
         }
     }
